activities and remove pfsense

+++ service/vm_service/vm_service.go
@@ -3,31 +3,27 @@ package vm_service
 import (
 	"fmt"
 	"go-deploy/models/dto/body"
-	"go-deploy/models/sys/vm"
+	vmModel "go-deploy/models/sys/vm"
 	"go-deploy/models/sys/vm/gpu"
 	"go-deploy/pkg/conf"
 	"go-deploy/service/vm_service/internal_service"
 	"log"
-
-	"go.mongodb.org/mongo-driver/bson"
 )
 
-func Create(vmID, name, sshPublicKey, owner string) error {
+func Create(vmID, owner string, vmCreate *body.VmCreate) error {
 	makeError := func(err error) error {
 		return fmt.Errorf("failed to create vm. details: %s", err)
 	}
 
-	err := vm.Create(vmID, name, sshPublicKey, owner, conf.Env.Manager)
+	params := vmModel.CreateParams{}
+	params.FromDTO(vmCreate)
+
+	err := vmModel.Create(vmID, params.Name, params.SshPublicKey, owner, conf.Env.Manager)
 	if err != nil {
 		return makeError(err)
 	}
 
-	csResult, err := internal_service.CreateCS(name, sshPublicKey)
-	if err != nil {
-		return makeError(err)
-	}
-
-	_, err = internal_service.CreatePfSense(name, csResult.PublicIpAddress.IpAddress)
+	_, err = internal_service.CreateCS(params.Name, params.SshPublicKey, params.Ports)
 	if err != nil {
 		return makeError(err)
 	}
@@ -35,8 +31,8 @@ func Create(vmID, name, sshPublicKey, owner string) error {
 	return nil
 }
 
-func GetByID(userID, vmID string, isAdmin bool) (*vm.VM, error) {
-	vm, err := vm.GetByID(vmID)
+func GetByID(userID, vmID string, isAdmin bool) (*vmModel.VM, error) {
+	vm, err := vmModel.GetByID(vmID)
 	if err != nil {
 		return nil, err
 	}
@@ -48,26 +44,20 @@ func GetByID(userID, vmID string, isAdmin bool) (*vm.VM, error) {
 	return vm, nil
 }
 
-func GetByOwnerID(ownerID string) ([]vm.VM, error) {
-	return vm.GetByOwnerID(ownerID)
+func GetByOwnerID(ownerID string) ([]vmModel.VM, error) {
+	return vmModel.GetByOwnerID(ownerID)
 }
 
-func GetAll() ([]vm.VM, error) {
-	return vm.GetAll()
+func GetAll() ([]vmModel.VM, error) {
+	return vmModel.GetAll()
 }
 
 func GetCount(userID string) (int, error) {
-	return vm.CountByOwnerID(userID)
+	return vmModel.CountByOwnerID(userID)
 }
 
-func Exists(name string) (bool, *vm.VM, error) {
-	return vm.Exists(name)
-}
-
-func MarkBeingDeleted(vmID string) error {
-	return vm.UpdateWithBsonByID(vmID, bson.D{{
-		"beingDeleted", true,
-	}})
+func Exists(name string) (bool, *vmModel.VM, error) {
+	return vmModel.Exists(name)
 }
 
 func Delete(name string) error {
@@ -75,9 +65,13 @@ func Delete(name string) error {
 		return fmt.Errorf("failed to delete vm. details: %s", err)
 	}
 
-	vm, err := vm.GetByName(name)
+	vm, err := vmModel.GetByName(name)
 	if err != nil {
 		return makeError(err)
+	}
+
+	if vm == nil {
+		return nil
 	}
 
 	err = internal_service.DeleteCS(name)
@@ -94,7 +88,28 @@ func Delete(name string) error {
 		return makeError(fmt.Errorf("failed to detach gpu from vm"))
 	}
 
-	err = internal_service.DeletePfSense(name)
+	return nil
+}
+
+func Update(vmID string, dtoVmUpdate *body.VmUpdate) error {
+	makeError := func(err error) error {
+		return fmt.Errorf("failed to update vm. details: %s", err)
+	}
+
+	vmUpdate := vmModel.UpdateParams{}
+	vmUpdate.FromDTO(dtoVmUpdate)
+
+	err := internal_service.UpdateCS(vmID, vmUpdate.Ports)
+	if err != nil {
+		return makeError(err)
+	}
+
+	err = vmModel.UpdateByID(vmID, &vmUpdate)
+	if err != nil {
+		return makeError(err)
+	}
+
+	err = vmModel.RemoveActivity(vmID, vmModel.ActivityBeingUpdated)
 	if err != nil {
 		return makeError(err)
 	}
@@ -102,39 +117,16 @@ func Delete(name string) error {
 	return nil
 }
 
-func Update(vmID string, dtoVmUpdate *body.VmUpdate) error {
-	vmUpdate := vm.VmUpdate{}
-
-	if dtoVmUpdate.Ports != nil {
-		var ports []vm.Port
-		for _, port := range *dtoVmUpdate.Ports {
-			ports = append(ports, vm.Port{
-				Name:     port.Name,
-				Port:     port.Port,
-				Protocol: port.Protocol,
-			})
-		}
-		vmUpdate.Ports = &ports
-	}
-
-	err := vm.UpdateByID(vmID, &vmUpdate)
-	if err != nil {
-		return fmt.Errorf("failed to update vm. details: %s", err)
-	}
-
-	return nil
-}
-
-func GetConnectionString(vm *vm.VM) (string, error) {
+func GetConnectionString(vm *vmModel.VM) (string, error) {
 	domainName := conf.Env.VM.ParentDomain
-	port := vm.Subsystems.PfSense.PortForwardingRuleMap["ssh"].ExternalPort
+	port := vm.Subsystems.CS.PortForwardingRuleMap["__ssh"].PublicPort
 
 	connectionString := fmt.Sprintf("ssh cloud@%s -p %d", domainName, port)
 
 	return connectionString, nil
 }
 
-func DoCommand(vm *vm.VM, command string) {
+func DoCommand(vm *vmModel.VM, command string) {
 	go func() {
 		csID := vm.Subsystems.CS.VM.ID
 		if csID == "" {
@@ -153,4 +145,72 @@ func DoCommand(vm *vm.VM, command string) {
 			return
 		}
 	}()
+}
+
+func StartActivity(vmID, activity string) (bool, string, error) {
+	canAdd, reason, err := CanAddActivity(vmID, activity)
+	if err != nil {
+		return false, "", err
+	}
+
+	if !canAdd {
+		return false, reason, nil
+	}
+
+	err = vmModel.AddActivity(vmID, activity)
+	if err != nil {
+		return false, "", err
+	}
+
+	return true, "", nil
+}
+
+func CanAddActivity(deploymentID, activity string) (bool, string, error) {
+	vm, err := vmModel.GetByID(deploymentID)
+	if err != nil {
+		return false, "", err
+	}
+
+	if vm == nil {
+		return false, "", err
+	}
+
+	switch activity {
+	case vmModel.ActivityBeingCreated:
+		return !vm.BeingDeleted(), "It is being deleted", nil
+	case vmModel.ActivityBeingDeleted:
+		return !vm.BeingCreated(), "It is being created", nil
+	case vmModel.ActivityBeingUpdated:
+		if vm.DoingOnOfActivities([]string{
+			vmModel.ActivityBeingCreated,
+			vmModel.ActivityBeingDeleted,
+			vmModel.ActivityBeingUpdated,
+			vmModel.ActivityAttachingGPU,
+			vmModel.ActivityDetachingGPU,
+		}) {
+			return false, "It should not be in creation, deletion, or updating, and should not be attaching or detaching a GPU", nil
+		}
+		return true, "", nil
+	case vmModel.ActivityAttachingGPU:
+		if vm.DoingOnOfActivities([]string{
+			vmModel.ActivityBeingCreated,
+			vmModel.ActivityBeingDeleted,
+			vmModel.ActivityAttachingGPU,
+			vmModel.ActivityDetachingGPU,
+		}) {
+			return false, "It should not be in creation or deletion, and should not be attaching or detaching a GPU", nil
+		}
+		return true, "", nil
+	case vmModel.ActivityDetachingGPU:
+		if vm.DoingOnOfActivities([]string{
+			vmModel.ActivityBeingCreated,
+			vmModel.ActivityBeingDeleted,
+			vmModel.ActivityAttachingGPU,
+		}) {
+			return false, "It should not be in creation or deletion, and should not be attaching a GPU", nil
+		}
+		return true, "", nil
+	}
+
+	return false, "", fmt.Errorf("unknown activity %s", activity)
 }
